+++ core/src/dird/storage.h
@@ -1,7 +1,7 @@
 /*
    BAREOS® - Backup Archiving REcovery Open Sourced
 
-   Copyright (C) 2018-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2018-2022 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -23,6 +23,7 @@
 #define BAREOS_DIRD_STORAGE_H_
 
 #include "ndmp/smc.h"
+#include "dird/dird.h"
 
 template <typename T> class alist;
 
@@ -82,6 +83,7 @@ slot_number_t GetBareosSlotNumberByElementAddress(
     smc_element_address_assignment* smc_elem_aa,
     slot_type_t slot_type,
     slot_number_t element_addr);
-
+bool IsSameStorageDaemon(StorageResource* read_storage,
+                         StorageResource* write_storage);
 } /* namespace directordaemon */
 #endif  // BAREOS_DIRD_STORAGE_H_