+++ CHANGELOG.md
@@ -119,6 +119,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - daemons: changed daemon CLI arguments parsing [PR #1187]
 - config parser: Refactor config reloading by using shared pointers instead of callbacks [PR #1151]
 - tests: remove unused config files [PR #1247]
+- dird: fix config reload and unit tests dependency issue [PR #1161]
 
 ### Deprecated
 - make_catalog_backup.pl is now a shell wrapper script which will be removed in version 23.